+++ crates/ruma-client-api/src/push/set_pushrule.rs
@@ -148,11 +148,11 @@ pub mod v3 {
             }
 
             let (scope, kind, rule_id): (RuleScope, RuleKind, String) =
-                serde::Deserialize::deserialize(serde::de::value::SeqDeserializer::<
+                Deserialize::deserialize(serde::de::value::SeqDeserializer::<
                     _,
                     serde::de::value::Error,
                 >::new(
-                    path_args.iter().map(::std::convert::AsRef::as_ref),
+                    path_args.iter().map(::std::convert::AsRef::as_ref)
                 ))?;
 
             let IncomingRequestQuery { before, after } =